+++ CHANGELOG.md
@@ -3,3 +3,9 @@ All notable changes to MemWord Project for Computer Programming II Class.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## 0.1 - 2018-11-07
+### Added
+- mainly dbmanager class
+- parsing csv files into db
+- querying data from db